requirements.txt
# === Core Deep Learning & Pose Detection ===
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.2.0         # YOLOv8 / YOLOv11 Pose 模型

# === Data Processing & Utilities ===
numpy>=1.26.0
pandas>=2.2.0
opencv-python>=4.9.0
tqdm>=4.66.0
scipy>=1.12.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
matplotlib>=3.8.0

# === AI Semantic Analysis ===
openai>=1.20.0             # GPT API
google-generativeai>=0.6.0 # Gemini API (若有整合)

# === Database Handling ===
mysql-connector-python>=9.0.0

# === Logging & System Management ===
colorlog>=6.8.0
psutil>=5.9.0

# === Optional but Useful ===
# 若使用 GPU 記得安裝 CUDA 對應版 torch
# pip install torch==2.2.0+cu121 torchvision==0.17.0+cu121 -f https://download.pytorch.org/whl/torch_stable.html


# === MMAction2 動作辨識套件 ===
openmim>=0.3.9
mmengine>=0.10.4
mmcv>=2.1.0
mmpose>=1.3.1
mmaction2>=1.2.0
